+++ tests/Kiota.Builder.Tests/Configuration/StructuredMimeTypesCollectionTests.cs
@@ -19,7 +19,7 @@ public sealed class StructuredMimeTypesCollectionTests
         Assert.Equal("application/json;q=1", mimeTypes.First(), StringComparer.OrdinalIgnoreCase);
         Assert.Equal("application/xml;q=0.8", mimeTypes.Last(), StringComparer.OrdinalIgnoreCase);
         Assert.DoesNotContain("application/atom+xml", mimeTypes);
-        Assert.Null(mimeTypes.GetPriority("application/atom+xml"));
+        Assert.Equal(0.8f, mimeTypes.GetPriority("application/atom+xml"));
         Assert.Equal(1, mimeTypes.GetPriority("application/json"));
         Assert.Equal(0.8f, mimeTypes.GetPriority("application/xml"));
     }
@@ -40,6 +40,8 @@ public sealed class StructuredMimeTypesCollectionTests
     [InlineData("application/json, application/xml, application/yaml", "application/json", "application/json;q=1")]
     [InlineData("application/json, application/xml, application/yaml", "application/json,text/plain", "application/json;q=1")]
     [InlineData("application/json, application/xml, application/yaml;q=0.8", "application/json,text/plain,application/yaml", "application/json;q=1,application/yaml;q=0.8")]
+    [InlineData("application/json, application/xml;q=0.9, application/yaml;q=0.8", "application/github+json", "application/github+json;q=1")]
+    [InlineData("application/vnd.topicus.keyhub+json;version=67, application/yaml;q=0.8", "application/vnd.topicus.keyhub+json;version=67", "application/vnd.topicus.keyhub+json;version=67;q=1")]
     public void MatchesAccept(string configuredTypes, string declaredTypes, string expectedTypes)
     {
         var mimeTypes = new StructuredMimeTypesCollection(configuredTypes.Split(',').Select(static x => x.Trim()));
@@ -52,6 +54,8 @@ public sealed class StructuredMimeTypesCollectionTests
     [InlineData("application/json, application/xml;q=0.9, application/yaml;q=0.8", "application/json", "application/json")]
     [InlineData("application/json, application/xml;q=0.9, application/yaml;q=0.8", "application/json,text/plain", "application/json")]
     [InlineData("application/json, application/xml;q=0.9, application/yaml;q=0.8", "application/json,text/plain,application/yaml", "application/json")]
+    [InlineData("application/json, application/xml;q=0.9, application/yaml;q=0.8", "application/github+json", "application/github+json")]
+    [InlineData("application/vnd.topicus.keyhub+json;version=67, application/yaml;q=0.8", "application/vnd.topicus.keyhub+json;version=67", "application/vnd.topicus.keyhub+json;version=67")]
     public void MatchesContentType(string configuredTypes, string declaredTypes, string expectedTypes)
     {
         var mimeTypes = new StructuredMimeTypesCollection(configuredTypes.Split(',').Select(static x => x.Trim()));